// packages/ic-management/src/candid/ic-management.d.ts
export interface read_canister_snapshot_metadata_response {
  globals: Array<
    | { f32: number }
    | { f64: number }
    | { i32: number }
    | { i64: bigint }
    | { v128: bigint }
  >;
  canister_version: bigint;
  source: { metadata_upload: any } | { taken_from_canister: any };
  certified_data: Uint8Array | number[];
  global_timer: [] | [{ active: bigint } | { inactive: null }];
  on_low_wasm_memory_hook_status:
    | []
    | [
        | { condition_not_satisfied: null }
        | { executed: null }
        | { ready: null },
      ];
  wasm_module_size: bigint;
  stable_memory_size: bigint;
  wasm_chunk_store: Array<{ hash: Uint8Array | number[] }>;
  taken_at_timestamp: bigint;
  wasm_memory_size: bigint;
}